+++ CHANGELOG.md
@@ -9,6 +9,10 @@
  - Fix `captureVals` on IE 7-8, make sure global flag works properly (#61, #62,
    #63)
 
+### Improvements
+
+ - `Route.interpolate()` accepts Numbers as segments. (#54)
+
 
 ## v0.9.0 (2012/05/28) ##